+++ src/main/kotlin/next/step/racing/domain/racing/Racing.kt
@@ -3,7 +3,7 @@ package next.step.racing.domain.racing
 import next.step.racing.domain.car.CarNames
 import next.step.racing.domain.car.CarRecord
 import next.step.racing.domain.car.Cars
-import next.step.racing.util.DrivingStrategy
+import next.step.racing.domain.car.DrivingStrategy
 
 object Racing {
