+++ src/test/kotlin/next/step/racing/domain/car/CarNamesTest.kt
@@ -1,9 +1,7 @@
-package next.step.racing.domain
+package next.step.racing.domain.car
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
-import next.step.racing.domain.car.CarName
-import next.step.racing.domain.car.CarNames
 
 class CarNamesTest : DescribeSpec({
