+++ src/test/kotlin/next/step/racing/domain/car/CarTest.kt
@@ -1,12 +1,8 @@
-package next.step.racing.domain
+package next.step.racing.domain.car
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
-import next.step.racing.domain.car.Car
-import next.step.racing.domain.car.CarName
-import next.step.racing.domain.car.CarPosition
-import next.step.racing.util.DrivingStrategy
 
 class CarTest : DescribeSpec({
