+++ src/test/kotlin/next/step/racing/domain/racing/RacingTest.kt
@@ -1,12 +1,9 @@
-package next.step.racing.view
+package next.step.racing.domain.racing
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.matchers.shouldBe
 import next.step.racing.domain.car.CarNames
 import next.step.racing.domain.car.CarRecord
-import next.step.racing.domain.racing.Racing
-import next.step.racing.domain.racing.RacingResult
-import next.step.racing.domain.racing.StepCount
 
 class RacingTest : DescribeSpec({
